Improve the API to add more sharing options
As a new method needed to be created to append new buttons to the share API, an opportunity to rewrite the core buttons was created. Now all the buttons, custom and core (facebook and twitter) are created the same way. This made possible to update the Facebook and Twitter sharing URLs, with the greatest benefit of dropping the Facebook appId and it's entire scripts. README and docs were update to reflect these changes.